+++ src/extensions/rules/rulesext.h
@@ -199,4 +199,10 @@ public:
      *  Is LandType Beach considered as "requires crushing" for passability purposes, as opposed to water?
      */
     bool IsBeachIsCrush;
+
+    /**
+     *  Defines for how many frames buildings do not get flames spawned on them on
+     *  damage state change after once catching fire.
+     */
+    int BuildingFlameSpawnBlockFrames;
 };