+++ docs/Whats-New.md
@@ -75,6 +75,7 @@ New:
 - Port to Syringe (by ZivDero)
 - Add customizable wake animations (by CCHyper, ZivDero)
 - Replace DirectDraw with SDL (by ZivDero, tomsons26, CCHyper)
+- Make it possible to prevent buildings from repeatedly catching fire when rapidly switching between damage stages (by Rampastring)
 
 
 Vinifera fixes: